requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
banks==2.2.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
filetype==1.2.0
frozenlist==1.8.0
fsspec==2025.9.0
google-auth==2.42.0
google-genai==1.46.0
greenlet==3.2.4
griffe==1.14.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
llama-index-core==0.14.6
llama-index-embeddings-google-genai==0.3.1
llama-index-instrumentation==0.4.2
llama-index-workflows==2.9.0
MarkupSafe==3.0.3
marshmallow==3.26.1
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
numpy==2.3.4
packaging==25.0
pillow==12.0.0
platformdirs==4.5.0
propcache==0.4.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
rsa==4.9.1
setuptools==80.9.0
sniffio==1.3.1
SQLAlchemy==2.0.44
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
wrapt==1.17.3
yarl==1.22.0
